+++ src/actions/utils/logging.ts
@@ -1,0 +1,4 @@
+"use server";
+import "server-only";
+
+export const serverLog = console.log;