+++ vitest.config.ts
@@ -1,0 +1,22 @@
+import react from "@vitejs/plugin-react";
+import path from "path";
+import tsconfigPaths from "vite-tsconfig-paths";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  plugins: [tsconfigPaths(), react()],
+  test: {
+    environment: "jsdom",
+    globals: true,
+    setupFiles: ["./src/test/setup.ts", "./src/test/setupElem.tsx"],
+    coverage: {
+      provider: "v8",
+      reporter: ["lcov", "html"],
+    },
+  },
+  resolve: {
+    alias: {
+      "@": path.resolve(__dirname, "./src"),
+    },
+  },
+});